Move statc methods for mastodon
