+++ src/mastodon.ts
@@ -1,3 +1,4 @@
+import { OAuth2 } from 'oauth'
 import MastodonAPI from './mastodon/api_client'
 import { ProxyConfig } from './proxy_config'
 import OAuth from './oauth'
@@ -70,7 +71,20 @@ export default class Mastodon implements MegalodonInterface {
     baseUrl = DEFAULT_URL,
     proxyConfig: ProxyConfig | false = false
   ): Promise<OAuth.AppData> {
-    return MastodonAPI.Client.registerApp(clientName, options, baseUrl, proxyConfig)
+    return this.createApp(clientName, options, baseUrl, proxyConfig).then(async appData => {
+      return this.generateAuthUrl(
+        appData.client_id,
+        appData.client_secret,
+        {
+          redirect_uri: appData.redirect_uri,
+          scope: options.scopes
+        },
+        baseUrl
+      ).then(url => {
+        appData.url = url
+        return appData
+      })
+    })
   }
 
   /**
@@ -91,7 +105,27 @@ export default class Mastodon implements MegalodonInterface {
     baseUrl = DEFAULT_URL,
     proxyConfig: ProxyConfig | false = false
   ): Promise<OAuth.AppData> {
-    return MastodonAPI.Client.createApp(client_name, options, baseUrl, proxyConfig)
+    const redirect_uris = options.redirect_uris || NO_REDIRECT
+    const scopes = options.scopes || DEFAULT_SCOPE
+
+    const params: {
+      client_name: string
+      redirect_uris: string
+      scopes: string
+      website?: string
+    } = {
+      client_name,
+      redirect_uris,
+      scopes
+    }
+    if (options.website) params.website = options.website
+
+    return MastodonAPI.Client.post<OAuth.AppDataFromServer>(
+      '/api/v1/apps',
+      params,
+      baseUrl,
+      proxyConfig
+    ).then((res: Response<OAuth.AppDataFromServer>) => OAuth.AppData.from(res.data))
   }
 
   /**
@@ -111,7 +145,16 @@ export default class Mastodon implements MegalodonInterface {
     },
     baseUrl = DEFAULT_URL
   ): Promise<string> {
-    return MastodonAPI.Client.generateAuthUrl(clientId, clientSecret, options, baseUrl)
+    return new Promise(resolve => {
+      const oauth = new OAuth2(clientId, clientSecret, baseUrl, undefined, '/oauth/token')
+      const url = oauth.getAuthorizeUrl({
+        redirect_uri: options.redirect_uri,
+        response_type: 'code',
+        client_id: clientId,
+        scope: options.scope
+      })
+      resolve(url)
+    })
   }
 
   // ======================================
@@ -145,7 +188,18 @@ export default class Mastodon implements MegalodonInterface {
     redirect_uri = NO_REDIRECT,
     proxyConfig: ProxyConfig | false = false
   ): Promise<OAuth.TokenData> {
-    return MastodonAPI.Client.fetchAccessToken(client_id, client_secret, code, baseUrl, redirect_uri, proxyConfig)
+    return MastodonAPI.Client.post<OAuth.TokenDataFromServer>(
+      '/oauth/token',
+      {
+        client_id,
+        client_secret,
+        code,
+        redirect_uri,
+        grant_type: 'authorization_code'
+      },
+      baseUrl,
+      proxyConfig
+    ).then((res: Response<OAuth.TokenDataFromServer>) => OAuth.TokenData.from(res.data))
   }
 
   /**
@@ -166,7 +220,17 @@ export default class Mastodon implements MegalodonInterface {
     baseUrl = DEFAULT_URL,
     proxyConfig: ProxyConfig | false = false
   ): Promise<OAuth.TokenData> {
-    return MastodonAPI.Client.refreshToken(client_id, client_secret, refresh_token, baseUrl, proxyConfig)
+    return MastodonAPI.Client.post<OAuth.TokenDataFromServer>(
+      '/oauth/token',
+      {
+        client_id,
+        client_secret,
+        refresh_token,
+        grant_type: 'refresh_token'
+      },
+      baseUrl,
+      proxyConfig
+    ).then((res: Response<OAuth.TokenDataFromServer>) => OAuth.TokenData.from(res.data))
   }
 
   /**
